apps/api/app/core/config.py
from pathlib import Path

from dotenv import load_dotenv
from pydantic_settings import BaseSettings, SettingsConfigDict

# Project root (apps/api/app/core/config.py -> 5 levels up: core->app->api->apps->root)
_ROOT = Path(__file__).resolve().parent.parent.parent.parent.parent
# API dir (apps/api)
_API_ROOT = _ROOT / "apps" / "api"

# Explicitly load .env files into os.environ (ensures CLERK_JWKS_URL etc. are available)
for env_path in (_ROOT / ".env", _API_ROOT / ".env"):
    if env_path.exists():
        load_dotenv(env_path, override=False)  # earlier files take precedence


class Settings(BaseSettings):
    # Load .env from project root and apps/api (api dir takes precedence for overrides)
    model_config = SettingsConfigDict(
        env_file=(_ROOT / ".env", _API_ROOT / ".env"),
        env_file_encoding="utf-8",
    )

    database_url: str = "postgresql+asyncpg://postgres:postgres@localhost:5432/interview_os"
    database_url_sync: str = "postgresql://postgres:postgres@localhost:5432/interview_os"

    # Demo gate
    demo_key: str | None = None  # DEMO_KEY env; if set, require x-demo-key header on non-public routes

    # Clerk auth (when set, Bearer token required; else fallback to demo_key + user_id)
    clerk_jwks_url: str | None = None  # CLERK_JWKS_URL e.g. https://<xxx>.clerk.accounts.dev/.well-known/jwks.json
    clerk_issuer: str | None = None  # CLERK_ISSUER override; if unset, derived from JWKS URL

    # CORS origins (comma-separated; when empty, uses default localhost list)
    cors_origins: str = ""  # CORS_ORIGINS e.g. http://localhost:3000,http://192.168.1.5:3000

    # AWS S3 (production)
    aws_region: str = "us-east-1"
    s3_bucket: str | None = None
    aws_access_key_id: str | None = None
    aws_secret_access_key: str | None = None

    # Hard limits (config via env)
    max_pdf_mb: int = 10  # MAX_PDF_MB
    max_pdf_pages: int = 20  # MAX_PDF_PAGES
    max_chunks_per_doc: int = 300  # MAX_CHUNKS_PER_DOC
    top_k_max: int = 8  # TOP_K_MAX
    max_completion_tokens: int = 500  # MAX_COMPLETION_TOKENS

    # Chunking (job description uses jd_chunking; these retained for potential generic docs)
    chunk_size: int = 512  # CHUNK_SIZE (legacy)
    min_chunk_chars: int = 25  # MIN_CHUNK_CHARS
    top_n_candidates: int = 50  # Fetch N by pgvector similarity before MMR
    mmr_lambda: float = 0.7  # MMR: lambda*sim(q,d) - (1-lambda)*max_sim(d,selected)

    # OpenAI embeddings
    openai_api_key: str | None = None  # OPENAI_API_KEY
    openai_embedding_model: str = "text-embedding-3-small"  # OPENAI_EMBEDDING_MODEL
    openai_embedding_dim: int = 1536  # OPENAI_EMBEDDING_DIM (must match DB vector column)

    # OpenAI chat (Q&A)
    openai_chat_model: str = "gpt-4o-mini"  # OPENAI_CHAT_MODEL


settings = Settings()
