apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
PyJWT>=2.8.0
cryptography>=42.0.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pgvector>=0.2.0
boto3>=1.34.0
openai>=1.0.0
pymupdf>=1.24.0

# Test
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
